pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-stats-cli"
version = "1.0.0"
description = "CLI tool for developer statistics from GitHub and local repos"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "gitpython>=3.1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
devstats = "dev_stats_cli.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
